Make exps seed. close #33

+++ migrations/20160915214407_applications.js
@@ -15,8 +15,8 @@ exports.up = function(knex) {
       .index();
     table.boolean('anonymous').defaultTo(false);
     table.decimal('overall_score');
-    table.decimal('edus_rating');
-    table.decimal('exps_rating');
+    table.integer('edus_rating');
+    table.integer('exps_rating');
   });
 };
